+++ backend/user/UserAPI/app/routers/seekers/education.py
@@ -18,8 +18,8 @@ router = APIRouter()
 async def user_seeker_education(
     db: Session = Depends(get_db), authorization: str = Header(...)
 ):
-    username = await get_current_user(authorization=authorization)
-    user_id = crud.seeker.base.get_userid_from_username(db=db, username=username)
+    user = await get_current_user(authorization=authorization)
+    user_id = user.get("user_id")
     user_education = crud.seeker.education.get_all(db=db, user_id=user_id)
     return user_education
